Rename FlexibleAlu to FlexibleFunctions.
