+++ coreblocks/fu/vector_unit/__init__.py
@@ -1,4 +1,4 @@
-from coreblocks.fu.vector_unit.flexible_alu import *  # noqa: F401
+from coreblocks.fu.vector_unit.flexible_functions import *  # noqa: F401
 from coreblocks.fu.vector_unit.vrf import *  # noqa: F401
 from coreblocks.fu.vector_unit.v_layouts import *  # noqa: F401
 from coreblocks.fu.vector_unit.v_register import *  # noqa: F401
